API Reference

BMPImages.BMPImages — Module

BMPImages provides the I/O support for Windows Bitmap (*.bmp) files.

APIs

read_bmp

read_bmp_header

write_bmp

add_bmp_format

BMPImages.BMPImageHeader — Type

BMPImageHeader

A mutable struct that holds the file header, information header, and color table for a bitmap.

The order and type of fields generally follows the original specifications.

BMPImages.add_bmp_format — Function

add_bmp_format()

Register format"BMP" to FileIO specifying BMPImages as the primary provider of the loader/saver.

Warning

This function forcefully rewrites the internal registry of FileIO.

BMPImages.read_bmp — Method

read_bmp(filepath::AbstractString; kwargs...)
read_bmp(io::IO; kwargs...)

Read a BMP image from the specified file or IO object.

Keyword arguments

not supported yet

Return types

Matrix{RGB{N0f8}}: default

Matrix{Gray{N0f8}}: for 4-bit or 8-bit indexed grayscale images

Matrix{Gray{Bool}}: for binary grayscale (black and white) images

BMPImages.write_bmp — Method

write_bmp(filepath::AbstractString, image; kwargs...)
write_bmp(io::IO, image; kwargs...)

Write a image as a BMP image to the specified file or IO object.

Keyword arguments

ppi::Real: pixel density in pixels per inch

expand_paletted::Bool: If true, the color table is not used regardless of the number of colors used. Default to false.